add daily goal api
