+++ src/features/daily-goal/api/index.ts
@@ -1,0 +1,3 @@
+export * from './addDailyGoal'
+export * from './fetchDailyGoals'
+export * from './updDailyGoal'
